更新StoreBuffer、CommitStage和ExecuteStage以支持64位地址和数据处理。
主要改动包括：
- 将相关字段和方法参数类型从32位更改为64位，确保兼容RV64I架构。
- 更新调试输出格式，统一使用64位格式化指令ID和地址，提升可读性。
- 优化Store-to-Load转发逻辑，确保在64位环境下正确处理数据。

此改动为RV64I指令集的支持奠定了基础，确保系统在64位模式下的稳定性和性能。

+++ include/cpu/ooo/stages/commit_stage.h
@@ -22,18 +22,18 @@ public:
 
 private:
     // 处理系统调用
-    void handle_ecall(CPUState& state, uint32_t instruction_pc);
+    void handle_ecall(CPUState& state, uint64_t instruction_pc);
     void handle_ebreak(CPUState& state);
     
     // 异常处理
-    void handle_exception(CPUState& state, const std::string& exception_msg, uint32_t pc);
+    void handle_exception(CPUState& state, const std::string& exception_msg, uint64_t pc);
     
     // 流水线刷新（用于跳转指令提交后）
     void flush_pipeline_after_commit(CPUState& state);
     void reset_execution_units(CPUState& state);
     
     // 调试辅助方法
-    void print_stage_activity(const std::string& activity, uint64_t cycle, uint32_t pc);
+    void print_stage_activity(const std::string& activity, uint64_t cycle, uint64_t pc);
 };
 
 } // namespace riscv
