+++ include/cpu/ooo/stages/execute_stage.h
@@ -26,13 +26,13 @@ private:
     void execute_instruction(ExecutionUnit& unit, DynamicInstPtr instruction, CPUState& state);
     void update_execution_units(CPUState& state);
     ExecutionUnit* get_available_unit(ExecutionUnitType type, CPUState& state);
-    bool execute_branch_operation(const DecodedInstruction& inst, uint32_t src1, uint32_t src2, CPUState& state);
-    void execute_store_operation(const DecodedInstruction& inst, uint32_t src1, uint32_t src2, CPUState& state);
+    bool execute_branch_operation(const DecodedInstruction& inst, uint64_t src1, uint64_t src2, CPUState& state);
+    void execute_store_operation(const DecodedInstruction& inst, uint64_t src1, uint64_t src2, CPUState& state);
     bool perform_load_execution(ExecutionUnit& unit, CPUState& state);
     
     // 分支预测相关
-    bool predict_branch(uint32_t pc);
-    void update_branch_predictor(uint32_t pc, bool taken);
+    bool predict_branch(uint64_t pc);
+    void update_branch_predictor(uint64_t pc, bool taken);
     void flush_pipeline(CPUState& state);
     void reset_execution_units(CPUState& state);
     
@@ -45,7 +45,7 @@ private:
     void complete_execution_unit(ExecutionUnit& unit, ExecutionUnitType unit_type, size_t unit_index, CPUState& state);
     
     // 调试辅助方法
-    void print_stage_activity(const std::string& activity, uint64_t cycle, uint32_t pc);
+    void print_stage_activity(const std::string& activity, uint64_t cycle, uint64_t pc);
 };
 
 } // namespace riscv
